requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Data Validation ──────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
asyncpg==0.29.0         # PostgreSQL async driver (production)
psycopg2-binary==2.9.9  # PostgreSQL sync (Alembic migrations)

# ── Cache / Session ───────────────────────────────────────────────────────────
redis[asyncio]==5.0.4

# ── LLM / Agent ───────────────────────────────────────────────────────────────
langchain==0.2.1
langchain-groq>=0.1.9
langchain-openai==0.1.7
langchain-core==0.2.3
langgraph==0.1.1
openai==1.29.0

# ── Voice – STT ───────────────────────────────────────────────────────────────
deepgram-sdk==3.4.0

# ── Voice – TTS ───────────────────────────────────────────────────────────────
elevenlabs==1.2.2

# ── Language Detection & Translation ─────────────────────────────────────────
langdetect==1.0.9
deep-translator==1.11.4

# ── Background Jobs ───────────────────────────────────────────────────────────
celery[redis]==5.4.0
kombu==5.3.5

# ── Observability ─────────────────────────────────────────────────────────────
structlog==24.1.0
python-json-logger==2.0.7

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==8.3.0       # Retry logic
pytz==2024.1
python-dateutil==2.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0         # async test client
